Remove GetSlice overload
It was inconsistent with the behaviour of how slice works,
meaning, whenever we saw \r\n it sliced the part behind it
and depending on whether the method handled crlf
the Range.End property was inclusive or exclusive.

So instead of having GetSlice to decide what to do with crlf,
I added the method GetSliceEnding that determines
which CRLF characters exists at the end of a slice.