Remove ReadAhead, MoveAhead and AbsoluteIndex and fix tests
Calls to ReadAhead and MoveAhead
were replaced with ReadChar and MoveNext respectively.

It was really difficult to follow the internal offset variable
and AbsoluteIndex property offered by ReadAhead and MoveAhead
and it created inconsistencies that in some circumstances
led to IndexOutOfRangeException So I decided to get rid of them.

Fix tests that relies on a generated random strings
that were filled with numbers that cause tests to fail
when they shouldn't.

Fix tests that were testing too many things.